\name{MRQoL-package}
\alias{MRQoL}
\docType{package}
\title{Minimal Clinically Important Difference and Response Shift Effect for Health-Related Quality of Life
}

\description{Using this package we can calculate the Minimal Clinically Important Difference by applying the Anchor-based method and the Response shift effect by applying the Then-Test method.
}

\details{
\tabular{ll}{
Package: \tab MCID\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-30\cr
License: GPL (>= 2.0)
}
}

\author{ Ahmad Ousmen
Maintainer: Ahmad Ousmen <ousmen.ahmad@hotmail.com>
}

\references{ 

Ann K Kvam, Finn Wisloff and Peter M Fayers, Minimal important differences and response shift in health related quality of life ; 
a longitudinal study in patients with multiple myeloma, Health and  Quality of Life  Outcomes, 2010;1477-7525-8-79.

T. S. Dabakuyo, F. Guillemin, T. Conroy, M. Velten, D. Jolly, M. Mercier, S. Causeret, J. Cuisenier, 
O. Graesslin, M. Gauthier, F. Bonnetain, Response shift effects on measuring post-operative quality of life
among breast cancer patients: a multicenter cohort study, Qual Life Res, 2012.

Zeinab Hamidou, Tienhan Sandrine Dabakuyo and Franck Bonnetain, Impact of response 
shift on longitudinal quality-of-life assessment in cancer clinical trials, Pharmacoeconomics Outcomes Res, 2011;11(5), 549-559.
}
